requirements.txt
pycrypto==2.6.1
PyYAML==6.0
